core: split checkpoint into capture/image submodules
